chore(web): add snapshot tests
